+++ src/paypal-checkout/paypal-checkout.js
@@ -256,6 +256,19 @@ PayPalCheckout.prototype._initialize = function (options) {
  */
 
 /**
+ * @typedef {object} PayPalCheckout~shippingOption
+ * @property {string} id A unique ID that identifies a payer-selected shipping option.
+ * @property {string} label A description that the payer sees, which helps them choose an appropriate shipping option. For example, `Free Shipping`, `USPS Priority Shipping`, `Expédition prioritaire USPS`, or `USPS yōuxiān fā huò`. Localize this description to the payer's locale.
+ * @property {boolean} selected If `selected = true` is specified as part of the API request it represents the shipping option that the payee/merchant expects to be pre-selected for the payer when they first view the shipping options within the PayPal checkout experience. As part of the response if a shipping option has `selected = true` it represents the shipping option that the payer selected during the course of checkout with PayPal. Only 1 `shippingOption` can be set to `selected = true`.
+ * @property {string} type The method by which the payer wants to get their items. The possible values are:
+ * * `SHIPPING` - The payer intends to receive the items at a specified address.
+ * * `PICKUP` - The payer intends to pick up the items at a specified address. For example, a store address.
+ * @property {object} amount The shipping cost for the selected option.
+ * @property {string} amount.currency The three-character ISO-4217 currency code. PayPal does not support all currencies.
+ * @property {string} amount.value The amount the shipping option will cost. Includes the specified number of digits after decimal separator for the ISO-4217 currency code.
+ */
+
+/**
  * Creates a PayPal payment ID or billing token using the given options. This is meant to be passed to PayPal's checkout.js library.
  * When a {@link callback} is defined, the function returns undefined and invokes the callback with the id to be used with the checkout.js library. Otherwise, it returns a Promise that resolves with the id.
  * @public
@@ -270,6 +283,7 @@ PayPalCheckout.prototype._initialize = function (options) {
  * @param {string} [options.currency] The currency code of the amount, such as 'USD'. Required when using the Checkout flow.
  * @param {string} [options.displayName] The merchant name displayed inside of the PayPal lightbox; defaults to the company name on your Braintree account
  * @param {string} [options.locale=en_US] Use this option to change the language, links, and terminology used in the PayPal flow. This locale will be used unless the buyer has set a preferred locale for their account. If an unsupported locale is supplied, a fallback locale (determined by buyer preference or browser data) will be used and no error will be thrown.
+ * @param {string} [options.vaultInitiatedCheckoutPaymentMethodToken] Use the payment method nonce representing a PayPal account with a Billing Agreement ID to create the payment and redirect the customer to select a new financial instrument. This option is only applicable to the `checkout` flow.
  *
  * Supported locales are:
  * `da_DK`,
@@ -296,6 +310,7 @@ PayPalCheckout.prototype._initialize = function (options) {
  * `zh_HK`,
  * and `zh_TW`.
  *
+ * @param {shippingOption[]} [options.shippingOptions] List of shipping options offered by the payee or merchant to the payer to ship or pick up their items.
  * @param {boolean} [options.enableShippingAddress=false] Returns a shipping address object in {@link PayPal#tokenize}.
  * @param {object} [options.shippingAddressOverride] Allows you to pass a shipping address you have already collected into the PayPal payment flow.
  * @param {string} options.shippingAddressOverride.line1 Street address.
@@ -328,6 +343,70 @@ PayPalCheckout.prototype._initialize = function (options) {
  *   },
  *   // Add other options, e.g. onApproved, onCancel, onError
  * }).render('#paypal-button');
+ *
+ * @example
+ * // shippingOptions are passed to createPayment. You can review the result from onAuthorize to determine which shipping option id was selected.
+ * ```javascript
+ * braintree.client.create({
+ *   authorization: 'authorization'
+ * }).then(function (clientInstance) {
+ *   return braintree.paypalCheckout.create({
+ *     client: clientInstance
+ *   });
+ * }).then(function (paypalCheckoutInstance) {
+ *   return paypal.Button.render({
+ *     env: 'production'
+ *
+ *     payment: function () {
+ *       return paypalCheckoutInstance.createPayment({
+ *         flow: 'checkout',
+ *         amount: '10.00',
+ *         currency: 'USD',
+ *         shippingOptions: [
+ *           {
+ *             id: 'UUID-9',
+ *             type: 'PICKUP',
+ *             label: 'Store Location Five',
+ *             selected: true,
+ *             amount: {
+ *               value: '1.00',
+ *               currency: 'USD'
+ *             }
+ *           },
+ *           {
+ *             id: 'shipping-speed-fast',
+ *             type: 'SHIPPING',
+ *             label: 'Fast Shipping',
+ *             selected: false,
+ *             amount: {
+ *               value: '1.00',
+ *               currency: 'USD'
+ *             }
+ *           },
+ *           {
+ *             id: 'shipping-speed-slow',
+ *             type: 'SHIPPING',
+ *             label: 'Slow Shipping',
+ *             selected: false,
+ *             amount: {
+ *               value: '1.00',
+ *               currency: 'USD'
+ *             }
+ *           }
+ *         ]
+ *       });
+ *     },
+ *
+ *     onAuthorize: function (data, actions) {
+ *       return paypalCheckoutInstance.tokenizePayment(data).then(function (payload) {
+ *         // Submit payload.nonce to your server
+ *       });
+ *     }
+ *   }, '#paypal-button');
+ * }).catch(function (err) {
+ *  console.error('Error!', err);
+ * });
+ * ```
  *
  * @returns {(Promise|void)} Returns a promise if no callback is provided.
  */
@@ -413,7 +492,8 @@ PayPalCheckout.prototype.tokenizePayment = function (tokenizeOptions) {
     ecToken: tokenizeOptions.paymentToken,
     billingToken: tokenizeOptions.billingToken,
     payerId: tokenizeOptions.payerID,
-    paymentId: tokenizeOptions.paymentID
+    paymentId: tokenizeOptions.paymentID,
+    shippingOptionsId: tokenizeOptions.shippingOptionsId
   };
 
   analytics.sendEvent(this._clientPromise, 'paypal-checkout.tokenization.started');
@@ -468,7 +548,8 @@ PayPalCheckout.prototype._formatPaymentResourceData = function (options) {
       noShipping: (!options.enableShippingAddress).toString(),
       addressOverride: options.shippingAddressEditable === false,
       landingPageType: options.landingPageType
-    }
+    },
+    shippingOptions: options.shippingOptions
   };
 
   if (options.flow === 'checkout') {
@@ -487,6 +568,14 @@ PayPalCheckout.prototype._formatPaymentResourceData = function (options) {
 
     if (options.hasOwnProperty('lineItems')) {
       paymentResource.lineItems = options.lineItems;
+    }
+
+    if (options.hasOwnProperty('vaultInitiatedCheckoutPaymentMethodToken')) {
+      paymentResource.vaultInitiatedCheckoutPaymentMethodToken = options.vaultInitiatedCheckoutPaymentMethodToken;
+    }
+
+    if (options.hasOwnProperty('shippingOptions')) {
+      paymentResource.shippingOptions = options.shippingOptions;
     }
 
     for (key in options.shippingAddressOverride) {
@@ -557,6 +646,10 @@ PayPalCheckout.prototype._formatTokenizePayload = function (response) {
 
   if (account.details && account.details.creditFinancingOffered) {
     payload.creditFinancingOffered = account.details.creditFinancingOffered;
+  }
+
+  if (account.details && account.details.shippingOptionId) {
+    payload.shippingOptionId = account.details.shippingOptionId;
   }
 
   return payload;